Реализует изменение счетчика видимых комментариев
