+++ js/full-image-open.js
@@ -1,11 +1,23 @@
 import { isEscapeKey } from './util.js';
-import { renderFullImage, socialCommentsList } from './full-image-render.js';
+import { socialCommentsList, socialCommentTotalCount } from './render-comments.js';
+import { renderFullImage } from './full-image-render.js';
 
 const bigPicture = document.querySelector('.big-picture');
 const bigPictureCancel = document.querySelector('.big-picture__cancel');
 const pictureLinks = document.querySelectorAll('.picture');
-const socialCommentCount = document.querySelector('.social__comment-count');
+const socialCommentShownCount = document.querySelector('.social__comment-shown-count');
+let socialCommentShownCountNumber = Number(socialCommentShownCount.textContent);
 const commentsLoader = document.querySelector('.comments-loader');
+
+const onCommentsLoaderClick = () => {
+  socialCommentShownCountNumber += 5;
+  socialCommentShownCount.textContent = socialCommentShownCountNumber;
+
+  if (socialCommentShownCountNumber > Number(socialCommentTotalCount.textContent)) {
+    socialCommentShownCount.textContent = socialCommentTotalCount.textContent;
+    commentsLoader.classList.add('hidden');
+  }
+};
 
 const onBigPictureKeyDown = (evt) => {
   if (isEscapeKey(evt)) {
@@ -16,11 +28,19 @@ const onBigPictureKeyDown = (evt) => {
 
 const openImageModal = (evt) => {
   bigPicture.classList.remove('hidden');
-  socialCommentCount.classList.add('hidden');
-  commentsLoader.classList.add('hidden');
   document.body.classList.add('modal-open');
   socialCommentsList.innerHTML = '';
+  socialCommentShownCount.textContent = 5;
+  socialCommentShownCountNumber = 5;
+
+  commentsLoader.classList.remove('hidden');
+  commentsLoader.addEventListener('click', onCommentsLoaderClick);
+
   renderFullImage(evt);
+  if (Number(socialCommentTotalCount.textContent) < 5) {
+    socialCommentShownCount.textContent = socialCommentTotalCount.textContent;
+    commentsLoader.classList.add('hidden');
+  }
 
   document.addEventListener('keydown', onBigPictureKeyDown);
 };
@@ -29,6 +49,7 @@ const closeImageModal = () => {
   bigPicture.classList.add('hidden');
   document.body.classList.add('modal-open');
 
+  commentsLoader.removeEventListener('click', onCommentsLoaderClick);
   document.removeEventListener('keydown', onBigPictureKeyDown);
 };
 
@@ -44,5 +65,3 @@ bigPictureCancel.addEventListener('click', (evt) => {
   evt.preventDefault();
   closeImageModal();
 });
-
-// Если начинаю перемещаться табом по изображениям, когда модальное окно уже открыто, и
